feat: Refactor Tron data models; remove unused fields and improve address handling

+++ chain-ingest/src/main/java/com/riskscoring/chainingest/client/dto/helius/HeliusTokenTransfer.java
@@ -1,9 +1,11 @@
 package com.riskscoring.chainingest.client.dto.helius;
 
+import java.math.BigDecimal;
+
 public record HeliusTokenTransfer(
         String fromUserAccount,
         String toUserAccount,
         String mint,
-        double tokenAmount
+        BigDecimal tokenAmount
 ) {
 }
